fix: attempts to address compilation issues in docker
